Initial support for combined status fetching
This WIP aims to provide much better performance on all supported devices
by requesting the data from all available modules at once.

The results are cached as usual, so update() needs to be awaited to get new values.
The initial update() will try to request as much data as possible, and simply ignore
the unavailable modules on consecutive requests.